requirements.txt
asgi-redis==0.8.3
asgiref==0.9
autobahn==0.12.1
channels==0.9.3
daphne==0.9.1
dj-database-url==0.4.0
Django==1.9.3
django-braces==1.9.0
djangorestframework==3.4.3
haikunator==1.0.1
msgpack-python==0.4.7
Pillow==3.3.0
psycopg2==2.6.1
redis==2.10.5
six==1.10.0
Twisted==15.5.0
txaio==2.2.1
zope.interface==4.1.3
